requirements.txt
matplotlib
opencv-python
numpy
pillow
pandas
PyQt5
